逆矩陣，向量，餘因子，伴隨矩陣完成。

+++ src/LinearAlgebra/Matrix/IdentityMatrix.kt
@@ -1,8 +1,8 @@
 package LinearAlgebra.Matrix
 import LinearAlgebra.Vector.Vector
-class IdentityMatrix(m:Int):OperableMatrix(MutableList<Vector>(m){Vector(MutableList<Float>(m){0f} )})
+class IdentityMatrix(m:Int):OperableMatrix(MutableList<Vector>(m){Vector(MutableList<Double>(m){0.0} )})
 {
     init {
-        for(i in 1 until m)this[i][i]=1f
+        for(i in 1 until m)this[i][i]=1.0
     }
 }
