+++ src/LinearAlgebra/Vector/Vector.kt
@@ -3,24 +3,24 @@ package LinearAlgebra.Vector
 
 open class Vector {
     var dim:Int
-    private var vector:MutableList<Float> = mutableListOf()
+    private var vector:MutableList<Double> = mutableListOf()
     var test = mutableListOf<Number>()
-    constructor(vararg ele:Float)
+    constructor(vararg ele:Double)
     {
         vector=ele.toMutableList()
         dim=ele.size
     }
     constructor(vararg ele:Int)
     {
-        val newMutableList = mutableListOf<Float>()
+        val newMutableList = mutableListOf<Double>()
         for(i in ele)
         {
-            newMutableList+=i.toFloat()
+            newMutableList+=i.toDouble()
         }
         vector=newMutableList
         dim=ele.size
     }
-    constructor(list: MutableList<Float>)
+    constructor(list: MutableList<Double>)
     {
         vector=list
         dim=list.size
@@ -28,15 +28,15 @@ open class Vector {
 
 
 
-    fun getVectorList():MutableList<Float>{return vector}
+    fun getVectorList():MutableList<Double>{return vector}
 
     override fun toString(): String {
         return vector.toString()
     }
 
-    operator fun get(m:Int):Float{return vector[m-1]}
+    operator fun get(m:Int):Double{return vector[m-1]}
 
-    operator fun set(i: Int, value: Float) { vector[i-1]=value }
+    operator fun set(i: Int, value: Double) { vector[i-1]=value }
 
 
 }
